Imputed member enrollment data created
I'm not a fan of putting imputed data with real data, but it is what it is. We added imputed rows to the monthly enrollment table for Medicaid based on proc and detail tables... or was it proc and header tables... anyway, we added imputed rows.

Also started on Truven things.